+++ app/page.tsx
@@ -1,6 +1,5 @@
 import Navbar from "../components/Navbar";
-import Hero from "../components/Hero";
-import Projects from "../components/Projects";
+import ResumeOpsHome from "../components/ResumeOpsHome";
 import EnterpriseExperience from "../components/EnterpriseExperience";
 import Skills from "../components/Skills";
 import CareerJourney from "../components/CareerJourney";
@@ -10,8 +9,7 @@ export default function Home() {
   return (
     <main id="overview" className="min-h-screen scroll-mt-24 bg-black text-white">
       <Navbar />
-      <Hero />
-      <Projects />
+      <ResumeOpsHome />
       <EnterpriseExperience />
       <Skills />
       <CareerJourney />