+++ next.config.ts
@@ -3,6 +3,7 @@ import type { NextConfig } from "next";
 const isProduction = process.env.NODE_ENV === "production";
 
 const nextConfig: NextConfig = {
+  distDir: ".next-v2",
   output: "export",
   trailingSlash: true,